Updated: code/data_utilities.py

+++ code/data_utilities.py
@@ -226,12 +226,15 @@ class PICTUREBCCTKDetectionDataset(Dataset):
             heatmap = transformed["mask"]
 
 
+        # Images post-processing
+        image = image.float()
+        
         # Keypoints post-processing
         keypoints = convert_albumentations_to_keypoints(keypoints)
         keypoints /= 512.
         keypoints = torch.from_numpy(keypoints).float()
 
         # Heatmap post processing
-        heatmap = torch.unsqueeze(heatmap, 0)
+        heatmap = torch.unsqueeze(heatmap, 0).float()
 
         return image, keypoints, heatmap
